+++ ThMouseX/Direct3D8.cpp
@@ -7,12 +7,12 @@
 #include <wrl/client.h>
 #include <comdef.h>
 #include <mutex>
+#include <cstdint>
 #include <imgui.h>
 #include "imgui_impl_win32.h"
 #include "imgui_impl_dx8.h"
 #include "ImGuiOverlay.h"
 
-#include "../Common/macro.h"
 #include "../Common/macro.h"
 #include "../Common/DataTypes.h"
 #include "../Common/MinHook.h"
@@ -167,8 +167,8 @@ namespace core::directx8 {
             return;
         }
 
-        auto vtable = *(DWORD**)pD3D.Get();
-        auto vtable2 = *(DWORD**)pDevice.Get();
+        auto vtable = *(uintptr_t**)pD3D.Get();
+        auto vtable2 = *(uintptr_t**)pDevice.Get();
 
         callbackstore::RegisterUninitializeCallback(TearDownCallback);
         callbackstore::RegisterClearMeasurementFlagsCallback(ClearMeasurementFlags);
